[core] Run PlatformIO in a subprocess via esphome.platformio_runner
Runs PlatformIO in a subprocess unconditionally instead of importing
platformio.__main__ in-process. A new esphome.platformio_runner module
is the subprocess entry point: it applies the incremental-rebuild and
download-retry patches before delegating to platformio.__main__.main().

This isolates PlatformIO's global state (including pioarduino's
sys.path stripping in penv_setup.py) from the ESPHome process, which
removes the need for the sys.path save/restore workaround.
